ZMQ-148: Change watermarks from int to uint64_t
git-svn-id: http://svn.fastmq.org/svn/zeromq/trunk@824 7451a71c-b6c8-49cc-8254-141fc5361860

+++ libzmq/bp_engine.cpp
@@ -111,7 +111,7 @@ zmq::engine_type_t zmq::bp_engine_t::type ()
     return engine_type_fd;
 }
 
-void zmq::bp_engine_t::get_watermarks (int *hwm_, int *lwm_)
+void zmq::bp_engine_t::get_watermarks (uint64_t *hwm_, uint64_t *lwm_)
 {
     *hwm_ = bp_hwm;
     *lwm_ = bp_lwm;
