+++ libzmq/out_engine.cpp
@@ -49,7 +49,7 @@ zmq::engine_type_t zmq::out_engine_t::type ()
     return engine_type_api;
 }
 
-void zmq::out_engine_t::get_watermarks (int *hwm_, int *lwm_)
+void zmq::out_engine_t::get_watermarks (uint64_t *hwm_, uint64_t *lwm_)
 {
     *hwm_ = 0;
     *lwm_ = 0;
