+++ libzmq/zmq/out_engine.hpp
@@ -37,7 +37,7 @@ namespace zmq
 
         //  i_engine implementation.
         engine_type_t type ();
-        void get_watermarks (int *hwm_, int *lwm_);
+        void get_watermarks (uint64_t *hwm_, uint64_t *lwm_);
         void process_command (const engine_command_t &command_);
 
     private:
